Add support for reading raw ultrasonic value
In certain situations it is useful to read the raw value from the
ultrasonic sensor. Especially when we want to have maximum precision.

Therefore, we add an option to read the value raw, before we remove some
of the information by dividing with 59 and rounding.

Hopefully, this can give students the possibility to navigate more
precisely.